Do not crash if JRebel is not on classpath
